requirements.txt
uvicorn
fastapi
aiofiles
pillow